Move verify function to verify_sandbox.hpp
